+++ src/automata/generations.cpp
@@ -18,7 +18,7 @@ Generations::Generations(std::string name, std::string rules)
     }
 
     // parse survive numbers
-    for (int i = 0; i < rules_arr[0].size(); i++) {
+    for (size_t i = 0; i < rules_arr[0].size(); i++) {
         int n = rules_arr[0][i] - '0';
         if (n < 0 || n > 9) {
             throw new std::runtime_error(
@@ -30,7 +30,7 @@ Generations::Generations(std::string name, std::string rules)
     }
 
     // parse birth numbers
-    for (int i = 0; i < rules_arr[1].size(); i++) {
+    for (size_t i = 0; i < rules_arr[1].size(); i++) {
         int n = rules_arr[1][i] - '0';
         if (n < 0 || n > 9) {
             throw new std::runtime_error(
@@ -57,28 +57,13 @@ std::pair<Board, bool> Generations::rewrite(const Board& board) {
     bool change_made = false;
     Board board_copy = board;
 
-    for (int row = 0; row < BOARD_ROWS; row++) {
-        for (int col = 0; col < BOARD_COLS; col++) {
+    for (size_t row = 0; row < BOARD_ROWS; row++) {
+        for (size_t col = 0; col < BOARD_COLS; col++) {
 
             // count neighbours
             uint8_t neighbour_count = get_neighbour_count(
                 board, NeighbourhoodType::Moore, row, col
             );
-            //uint8_t neighbour_count = 0;
-            //for (int row_offset = -1; row_offset <= 1; row_offset++) {
-            //    for (int col_offset = -1; col_offset <= 1; col_offset++) {
-            //        if (row_offset == 0 && col_offset == 0) {
-            //            continue;
-            //        }
-
-            //        auto neighbour_row = modulo(row + row_offset, BOARD_ROWS);
-            //        auto neighbour_col = modulo(col + col_offset, BOARD_COLS);
-
-            //        if (board[neighbour_row][neighbour_col] == 1) {
-            //            neighbour_count++;
-            //        }
-            //    }
-            //}
 
             if (board[row][col] == 0) {
                 if (contains(birth_numbers, neighbour_count)) {
